Removed jsonschema module
Because after more than two years of using it as a non installed module, I discovered that it can't be used if the module is also installed on the system. This is because internally `jsonschema` module imports its own modules without using relative imports (¬¬). Yet another reason why I abhor using third-party tools!

I'm going to stick to using the installed `jsonschema` module and if it isn't, f*ck validation and move the f*ck on!